# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "aptpay App API",
    "description": "Find out how your APIs work",
    "license": {
      "name": "tejas"
    }
  },
  "host": "uataptpay.moreyeahs.in",
  "basePath": "/api",
...
  "schemes": ["https", "http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "security": [
    {
      "APIKeyHeader": []
    }
  ],
  "paths": {
...
    "/crossborder/createTransaction": {
      "post": {
        "summary": "createTransaction",
        "tags": ["crossborder"],
        "description": "createTransaction",
        "parameters": [
          {
            "name": "createTransaction",
            "in": "body",
            "description": "createTransaction",
            "schema": {
              "$ref": "#/definitions/createTransaction"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/createTransactionResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/createTransactionErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/getCountryServices": {
      "get": {
        "summary": "getCountryServices",
        "tags": ["crossborder"],
        "description": "getCountryServices",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getCountryServicesResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getCountryServicesErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/getSorcesOfFunds": {
      "get": {
        "summary": "getSorcesOfFunds",
        "tags": ["crossborder"],
        "description": "getSorcesOfFunds",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getSorcesOfFundsResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getSorcesOfFundsErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/type-of-ids/{country}": {
      "get": {
        "summary": "getSenderTypeOfIDs",
        "tags": ["crossborder"],
        "description": "getSenderTypeOfIDs",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "description": "type-of-ids/{country}",
            "schema": {
              "type": "string"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/type-of-ids/{country}Response"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/type-of-ids/{country}ErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/type-of-ids-receiver/{country}": {
      "get": {
        "summary": "getReceiverTypeOfIDs",
        "tags": ["crossborder"],
        "description": "getReceiverTypeOfIDs",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "description": "createTransaction",
            "schema": {
              "type": "string"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/type-of-ids-receiver/:countryResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/type-of-ids-receiver/:countryErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/type-of-accounts/{country}": {
      "get": {
        "summary": "getAccountTypes",
        "tags": ["crossborder"],
        "description": "getAccountTypes",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "description": "createTransaction",
            "schema": {
              "type": "string"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/type-of-accounts/:countryResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/type-of-accounts/:countryErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/cities/{country}": {
      "get": {
        "summary": "getCities",
        "tags": ["crossborder"],
        "description": "getCities",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "description": "createTransaction",
            "schema": {
              "type": "string"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/cities/:country"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/cities/:country"
            }
          }
        }
      }
    },
    "/crossborder/banks/{country}": {
      "get": {
        "summary": "getBanks",
        "tags": ["crossborder"],
        "description": "getBanks",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "description": "createTransaction",
            "schema": {
              "type": "string"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/banks/:countryResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/banks/:countryErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/bank/{bankId}/branches": {
      "get": {
        "summary": "getBankBranches",
        "tags": ["crossborder"],
        "description": "getBankBranches",
        "parameters": [
          {
            "name": "bankId",
            "in": "path",
            "description": "createTransaction",
            "schema": {
              "type": "string"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/bank/:bankId/branchesResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/bank/:bankId/branchesErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/purposes/{country}": {
      "get": {
        "summary": "getPurposes",
        "tags": ["crossborder"],
        "description": "getPurposes",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "description": "createTransaction",
            "schema": {
              "type": "string"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/purposes/:countryResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/purposes/:countryErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/getPayers": {
      "post": {
        "summary": "getPayers",
        "tags": ["crossborder"],
        "description": "getPayers",
        "parameters": [
          {
            "name": "getPayers",
            "in": "body",
            "description": "getPayers",
            "schema": {
              "$ref": "#/definitions/getPayers"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getPayersResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getPayersErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/calculate": {
      "post": {
        "summary": "calculate",
        "tags": ["crossborder"],
        "description": "calculate",
        "parameters": [
          {
            "name": "calculate",
            "in": "body",
            "description": "calculate",
            "schema": {
              "$ref": "#/definitions/calculate"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/calculateResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/calculateErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/exchangeRate": {
      "post": {
        "summary": "exchangeRate",
        "tags": ["crossborder"],
        "description": "exchangeRate",
        "parameters": [
          {
            "name": "exchangeRate",
            "in": "body",
            "description": "exchangeRate",
            "schema": {
              "$ref": "#/definitions/exchangeRate"
            }
          }
        ],

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/exchangeRateResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/exchangeRateErrorResponse"
            }
          }
        }
      }
    },
    "/crossborder/newRecipient": {
      "post": {
        "summary": "new Recipient",
        "tags": ["crossborder"],
        "description": "new Recipient",
        "parameters": [
          {
            "name": "newRecipient",
            "in": "body",
            "description": "new Recipient",
            "schema": {
              "$ref": "#/definitions/newRecipient"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "new Recipient Success",
            "schema": {
              "$ref": "#/definitions/newRecipientResponse"
            }
          },
          "200": {
            "description": "new Recipient failed",
            "schema": {
              "$ref": "#/definitions/newRecipientErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createTransaction": {
      "type": "object",
      "properties": {
        "receiver": {
          "properties": {
            "payeeId": {
              "type": "number"
            }
          }
        },
        "transaction": {
          "properties": {
            "amount": {
              "type": "number"
            },
            "paymentMode": {
              "type": "string"
            },
            "sourceCurrency": {
              "type": "string"
            },
            "receiveCurrency": {
              "type": "string"
            },
            "account": {
              "type": "string"
            },
            "branch": {
              "type": "string"
            },
            "accountType": {
              "type": "string"
            },
            "purpose": {
              "type": "string"
            },
            "sourceOfFunds": {
              "type": "string"
            }
          }
        }
      }
    },
    "createTransactionResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "properties": {
            "payeeId": {
              "type": "number"
            },
            "amount": {
              "type": "number"
            },
            "cardHash": {
              "type": "string"
            }
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
...
  }
}